Finish-success

+++ fe/src/routes/home/blogs/+page.svelte
@@ -4,6 +4,8 @@
     import { goto } from '$app/navigation';
     import { blogs } from '../../../stores/blogs'
 
+    export let data;
+
     let notyf;
 
     onMount(() => {
@@ -14,22 +16,22 @@
 		});
 
         fetch("/api/showAllBlogs")
-            .then((v) => {
-                if(!v.ok) {
-                    throw new Error("服务端异常");
-                }
-                return v.json(); // 已经将 blogs 解析
-            })
-            .then((v) => {
-                if(v && v.status !== 0) {
-                    notyf.error(v.msg);
-                    return;
-                }
-                blogs.set(v.blogs);
-            })
-            .catch((e) => {
-                notyf.error(e.message);
-            });
+        .then((v) => {
+            if(!v.ok) {
+                throw new Error("服务端异常");
+            }
+            return v.json(); // 已经将 blogs 解析
+        })
+        .then((v) => {
+            if(v && v.status !== 0) {
+                notyf.error(v.msg);
+                return;
+            }
+            blogs.set(v.blogs);
+        })
+        .catch((e) => {
+            notyf.error(e.message);
+        });
     });
 
     let deleteBlog = (id) => {
@@ -42,20 +44,24 @@
             })
             .then((v) => {
                 if(v && v.status !== 0) {
-                    notyf.error("v.msg");
+                    notyf.error(v.msg);
                     return;
                 }
                 notyf.success(v.msg);  
                 blogs.set($blogs.filter(blog => blog.id !== id));
-            })
+            })  
             .catch((e) => {
                 notyf.error(e.message);
             });            
     };
 </script>
 
-<h2>我的文章</h2>
-<button on:click={() => { goto("/home/blogs/create") }} class="create-btn">点击创作</button>
+{#if data.role === 'user'}
+    <h2>我的文章</h2>
+    <button on:click={() => { goto("/home/blogs/create") }} class="create-btn">点击创作</button>    
+{:else}
+    <h2>所有文章</h2>
+{/if}
 {#if !$blogs?.length}
     <p>暂无文章</p>
 {:else}
@@ -63,10 +69,15 @@
         {#each $blogs as blog (blog.id)}
             <li class="blog-item" id={blog.id}>
                 <div class="blog-title">{blog.title}</div>
-                <div class="blog-content">...</div>
+                {#if data.role === 'admin'}
+                    <div class="blog-author">{blog.author}</div>
+                {/if}
+                <div class="blog-content">{@html blog.content}</div>
                 <div class="blog-actions">
                     <button on:click={() => { goto(`/home/blogs/show?id=${blog.id}`) }} class="show-btn">阅读</button>
-                    <button on:click={() => { goto(`/home/blogs/edit?id=${blog.id}`) }} class="edit-btn">编辑</button>
+                    {#if data.role === 'user'}
+                        <button on:click={() => { goto(`/home/blogs/edit?id=${blog.id}`) }} class="edit-btn">编辑</button>
+                    {/if}
                     <button on:click={() => deleteBlog(blog.id)} class="delete-btn">删除</button>
                 </div>
             </li>
@@ -93,10 +104,19 @@
         margin-bottom: 10px;
     }
 
+    .blog-author {
+        font-size: 18px;
+        font-weight: bold;
+        color: #333;
+        margin-bottom: 10px;
+    }
+
     .blog-content {
         font-size: 14px;
         color: #555;
         line-height: 1.6;
+        height: 50px;
+        overflow: hidden;
     }
 
     .blog-actions {
